arbitrary compile time code execution!!

+++ src/SyntaxNodes/RawData.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class RawData implements Node {
-    public List<Long> data = new ArrayList<>();
+    public List<Integer> data = new ArrayList<>();
 }
